requirements.lock.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
filetype==1.2.0
google-api-core==2.30.1
google-auth==2.49.1
google-genai==1.70.0
googleapis-common-protos==1.73.1
h11==0.16.0
hiredis==3.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-anthropic==1.4.0
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-openai==1.1.9
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
openai==1.109.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
regex==2026.3.32
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==1.0.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
